+++ src/main.py
@@ -6,7 +6,7 @@ from enum import Enum
 
 class MDFile:
     def __init__(self):
-        self.classes: list[MDClass] = []
+        self.classes: list[MDClass] = {}
 
 
 class MDClass:
@@ -44,7 +44,7 @@ class MDProperty:
 
 class PYFile:
     def __init__(self):
-        self.classes = []
+        self.classes = {}
 
 
 class PYClass:
@@ -61,6 +61,9 @@ class PYMethod:
     def __init__(self, name: str, params):
         self.name: str = name
         self.params = params
+
+    def __str__(self):
+        return f"{self.name}"
 
 
 class PYProperty:
@@ -82,6 +85,8 @@ class Parsnip:
 
         self.attr_store_type = AttrType.properties;
 
+        self.current_class = None
+
         md_data = self.parse_md_file("data/test.md")
         self.store_md_data(md_data)
         
@@ -92,7 +97,7 @@ class Parsnip:
         with open(filepath, "rb") as md:
             data = md.read();  # read in .md file
             # :-1 to remove trailing '\0'
-            parsed_data = subprocess.check_output(["cabal", "exec", "parsnip", data, "md"], text=True)[:-1] 
+            parsed_data = subprocess.check_output(["cabal", "exec", "parsnip", data, "md"], text=True)[:-1]
             md_data = json.loads(parsed_data)
 
             return md_data
@@ -100,7 +105,7 @@ class Parsnip:
     def parse_py_file(self, filepath: str) -> list[str]:
         with open(filepath, "rb") as py:
             data = py.read()
-            parsed_data = subprocess.check_output(["cabal", "exec", "parsnip", data, "py"], text=True)
+            parsed_data = subprocess.check_output(["cabal", "exec", "parsnip", data, "py"], text=True)[:-1]
             py_data = json.loads(parsed_data)
 
             return py_data
@@ -112,7 +117,9 @@ class Parsnip:
             rest = items[1:]           # store remaining values
 
             if item_type in ["Class", "Subclass"]:
-                self.md_file.classes.append(MDClass(*rest))  # store the remaining data in a class [name, description, ?superclass]
+                class_ = MDClass(*rest)
+                self.md_file.classes[rest[0]] = class_  # store the remaining data in a class [name, description, ?superclass]
+                self.current_class = class_
 
             elif item_type == "Subheader":
                 # checking whether we are about to read in properties or methods
@@ -127,9 +134,9 @@ class Parsnip:
             elif (item_type == "CodeSnippet"):
                 if self.attr_store_type == AttrType.properties:
                     # add to most recent class
-                    self.md_file.classes[-1].properties.append(MDProperty(*rest))
+                    self.current_class.properties.append(MDProperty(*rest))
                 elif self.attr_store_type == AttrType.methods:
-                    self.md_file.classes[-1].methods.append(MDMethod(*rest))
+                    self.current_class.methods.append(MDMethod(*rest))
 
     def store_py_data(self, data: list[str]):
         for line in data:
@@ -138,21 +145,52 @@ class Parsnip:
             rest = items[1:]
 
             if item_type == "Class":
-                self.py_file.classes.append(PYClass(*rest))
+                class_ = PYClass(*rest)
+                self.py_file.classes[rest[0]] = class_
+                self.current_class = class_
 
             if item_type == "Function":
                 name = rest[0]
                 params = rest[2:-1]
-                self.py_file.classes[-1].methods.append(PYMethod(name, params))
+                self.current_class.methods.append(PYMethod(name, params))
 
             if item_type == "Property":
-                self.py_file.classes[-1].properties.append(PYProperty(*rest))
+                self.current_class.properties.append(PYProperty(*rest))
+
+
+def red(s: str): return f"\x1b[;31m{s}\x1b[;39m"
+def green(s: str): return f"\x1b[;32m{s}\x1b[;39m"
+def yellow(s: str): return f"\x1b[;33m{s}\x1b[;39m"
+def blue(s: str): return f"\x1b[;34m{s}\x1b[;39m"
+def pink(s: str): return f"\x1b[;35m{s}\x1b[;39m"
+def cyan(s: str): return f"\x1b[;36m{s}\x1b[;39m"
+
+def red_bg(s: str): return f"\x1b[0;41m{s}\x1b[0;39m"
+def green_bg(s: str): return f"\x1b[0;42m{s}\x1b[0;39m"
+def yellow_bg(s: str): return f"\x1b[0;43m{s}\x1b[0;39m"
+def blue_bg(s: str): return f"\x1b[0;44m{s}\x1b[0;39m"
+def pink_bg(s: str): return f"\x1b[0;45m{s}\x1b[0;39m"
+def cyan_bg(s: str): return f"\x1b[0;46m{s}\x1b[0;39m"
+
+def bold(s: str): return f"\x1b[1;29m{s}\x1b[0;39m"
+def dim(s: str): return f"\x1b[2;29m{s}\x1b[0;39m"
+def italic(s: str): return f"\x1b[3;29m{s}\x1b[0;39m"
+def underline(s: str): return f"\x1b[4;29m{s}\x1b[0;39m"
 
 
 
 if __name__ == "__main__":
     app = Parsnip()
 
-    for c in app.py_file.classes:
-        print(c)
+    for class_name, c in app.py_file.classes.items():
+        if class_name not in app.md_file.classes:
+            print(f"[parsnip]: missing class - {red(dim(c))}")
+            continue
+
+        md_methods = [m.name for m in app.md_file.classes[class_name].methods]
+
+        for m in c.methods:
+            if m.name not in md_methods and m.name != "__init__":
+                print(f"[parsnip]: missing method - {red(c)}.{yellow(dim(m))}")
+